split topic tests into different files
